feat: move file service @service/waiting
